+++ Count/src/com/mzlabs/count/divideandconquer/RowCannonNode.java
@@ -82,7 +82,7 @@ final class RowCannonNode implements NonNegativeIntegralCounter {
 	}
 
 	@Override
-	public int cacheSize() {
+	public long cacheSize() {
 		return underlying.cacheSize();
 	}
